fix: cellphone, cnpj and currency parsers and formatters
Closes: #73

What?
Fix cellphone, CNPJ and currency formatters and parsers

Why?
Cellphone field on a user requires to be 15 characters long and must include a '+'
CNPJ field on a company requires to be 14 characters long and cannot contain symbols
Currency parser changed to always return a numeric value to maintain data consistency

How to test it?
yarn test src/utils